20241106-multi-agents-langgraph/career-assistant/requirements.txt
pypdf
langchain
langgraph
langchain-openai
python-dotenv
linkedin-api
streamlit
langsmith
streamlit_chat
streamlit_pills
langchain_community
firecrawl-py
pymupdf
streamlit-analytics2
python-docx
asgiref